AST: Split off Requirement.cpp from GenericSignature.cpp
